Handle slashes in the branch name for upstream branch folding
If the branch name includes `/` character(s) the upstream folding fails
as we only compare a sub-part of the branch with the upstream branch
which still has the `/` character(s) in it.

Here we manipulate the `parts` array prior to the compare so that we can
include `/`'s when they did exist in the remote branch name.

`${x[@]:1}` gets all but the first element of the `$x` array
`${(j:/:)x}` joins all elements of `$x` with `/`